+++ WindowsSandboxTools.psm1
@@ -1,0 +1,4 @@
+
+#dot source the module functions
+Get-ChildItem -path $psscriptroot\functions\*.ps1 |
+Foreach-Object { . $_.fullname }